+++ CHANGELOG.md
@@ -9,7 +9,9 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
+- App: Fix invalid URLs when uploading assets.
 - Server: Allow CLI to update app block themes if it has the `apps:write` scope.
+- Server: Fix issue with invalid role validation for pages with sub pages.
 
 ## [0.11.6] - 2020-03-05